All classes defined in the package of `Rcwl` and the class constructor functions.

InputArrayParam: Parameters for array inputs. To specify an array parameter, the array definition is nested under the type field with 'type: array' and items defining the valid data types that may appear in the array.

InputParam: parameter for a command line tool.

InputParamList: A list of `InputParam` objects.

OutputArrayParam: Parameters for array outputs.

OutputParam: An output parameter for a Command Line Tool.

OutputParamList: A list of `InputParam` objects.

stepInParam: The input parameter of a workflow step.

stepInParamList: A list of `stepInParam` objects.

cwlStepList: A list of `cwlStep` objects.

InputArrayParam(
  label = "",
  type = "array",
  items = character(),
  prefix = "",
  separate = TRUE,
  itemSeparator = character(),
  valueFrom = character()
)

InputParam(
  id,
  label = "",
  type = "string",
  doc = character(),
  secondaryFiles = character(),
  streamable = logical(),
  format = character(),
  loadListing = character(),
  loadContents = logical(),
  position = 0L,
  prefix = "",
  separate = TRUE,
  itemSeparator = character(),
  valueFrom = character(),
  shellQuote = logical(),
  default = character(),
  value = character()
)

InputParamList(...)

OutputArrayParam(
  label = character(),
  type = "array",
  items = character(),
  glob = character(),
  loadContents = logical(),
  outputEval = character()
)

OutputParam(
  id = "output",
  label = character(),
  doc = character(),
  type = "stdout",
  format = character(),
  secondaryFiles = character(),
  streamable = logical(),
  glob = character(),
  loadContents = logical(),
  outputEval = character(),
  outputSource = character()
)

OutputParamList(out = OutputParam(), ...)

stepInParam(
  id,
  source = character(),
  linkMerge = character(),
  default = character(),
  valueFrom = character()
)

stepInParamList(...)

cwlStepList(...)

Arguments

label
A short, human-readable label of this object.

type
Specify valid types of data that may be assigned to this parameter.

items
Defines the type of the array elements.

prefix
Command line prefix to add before the value.

separate
If true (default), then the prefix and value must be added as separate command line arguments; if false, prefix and value must be concatenated into a single command line argument.

itemSeparator
Join the array elements into a single string with the elements separated by by itemSeparator.

valueFrom
value from string or expression.

id
A unique identifier for this workflow input parameter.

doc
A documentation string for this object, or an array of strings which should be concatenated.

secondaryFiles
Provides a pattern or expression specifying files or directories. Only valid when type: File or is an array of items: File.

streamable
A value of true indicates that the file is read or written sequentially without seeking. Only valid when type: File or is an array of items: File.

format
Only valid when type: File or is an array of items: File. This is the file format that will be assigned to the output File object.

loadListing
Only valid when type: Directory or is an array of items: Directory.

loadContents
Read text from globbed file.

position
The position for this parameter.

shellQuote
If ShellCommandRequirement is in the requirements for the current command, this controls whether the value is quoted on the command line (default is true).

default
The default value for this parameter to use if either there is no source field, or the value produced by the source is null.

value
Assigned value for this parameter

...
A list of `cwlStep` objects.

glob
Pattern to find files relative to the output directory.

outputEval
Evaluate an expression to generate the output value.

outputSource
Specifies one or more workflow parameters that supply the value of to the output parameter.

out
The default stdout parameter.

source
Specifies one or more workflow parameters that will provide input to the underlying step parameter.

linkMerge
The method to use to merge multiple inbound links into a single array.

Value

InputArrayParam: An object of class `InputArrayParam`.

An object of class `InputParam`.

InputParamList: An object of class `InputParamList`.

An object of class `OutputArrayParam`.

OutputParam: An object of class `OutputParam`.

OutputParamList: An object of class `OutputParamList`.

stepInParam: An object of class `stepInParam`.

An object of class `stepInParamList`.

cwlStepList: An object of class `cwlStepList`.

Details

More details of `InputArrayParam`, see: https://www.commonwl.org/v1.0/CommandLineTool.html#CommandInputArraySchema

More details for `InputParam`, see: https://www.commonwl.org/v1.0/CommandLineTool.html#CommandInputParameter

More details for `OutputArrayParam`, see: https://www.commonwl.org/v1.0/CommandLineTool.html#CommandOutputArraySchema

More details for `OutputParam`, see: https://www.commonwl.org/v1.0/CommandLineTool.html#CommandOutputParameter

More details for `stepInParam`, see: https://www.commonwl.org/v1.0/Workflow.html#WorkflowStepInput

Examples

InputArrayParam(items = "string", prefix="-B=", separate = FALSE)

#> An object of class "InputArrayParam"
#> Slot "label":
#> [1] ""
#> 
#> Slot "type":
#> [1] "array"
#> 
#> Slot "items":
#> [1] "string"
#> 
#> Slot "inputBinding":
#> $prefix
#> [1] "-B="
#> 
#> $separate
#> [1] FALSE
#> 
#> $itemSeparator
#> character(0)
#> 
#> $valueFrom
#> character(0)
#> 
#> 

input1 <- InputParam(id = "sth")
InputParamList(input1)

#> inputs:
#>   sth (string):  

OutputParam(id = "b", type = OutputArrayParam(items = "File"), glob = "*.txt")

#> - type:
#>     type: array
#>     items: File
#>     outputBinding: {}
#>   outputBinding:
#>     glob: '*.txt'

o1 <- OutputParam(id = "file", type = "File", glob = "*.txt")
o1

#> - type: File
#>   outputBinding:
#>     glob: '*.txt'

o1 <- OutputParam(id = "file", type = "File", glob = "*.txt")
OutputParamList(o1)

#> outputs:
#> file:
#>   type: File
#>   outputBinding:
#>     glob: '*.txt'

s1 <- stepInParam(id = "s1")

s1 <- stepInParam(id = "s1")
stepInParamList(s1)

#> stepInParamList of length 1
#> names(1): s1

s1 <- cwlStep(id = "s1")
cwlStepList(s1)

#> steps:
#> s1:
#>   run: s1.cwl
#>   out:
#>   - output