InputArrayParam

Parameters for array inputs. To specify an array parameter, the array definition is nested under the type field with 'type: array' and items defining the valid data types that may appear in the array. More details: https://www.commonwl.org/v1.0/CommandLineTool.html#CommandInputArraySchema

InputArrayParam(
  label = "",
  type = "array",
  items = character(),
  prefix = "",
  separate = TRUE,
  itemSeparator = character(),
  valueFrom = character()
)

Arguments

label
A short description for this object

type
Must be "array".

items
Defines the type of the array elements.

prefix
Command line prefix to add before the value.

separate
If true (default), then the prefix and value must be added as separate command line arguments; if false, prefix and value must be concatenated into a single command line argument.

itemSeparator
Join the array elements into a single string with separator.

valueFrom
String or Expression.

Value

An object of class `InputArrayParam`.

Examples

InputArrayParam(items = "string", prefix="-B=", separate = FALSE)

#> An object of class "InputArrayParam"
#> Slot "label":
#> [1] ""
#> 
#> Slot "type":
#> [1] "array"
#> 
#> Slot "items":
#> [1] "string"
#> 
#> Slot "inputBinding":
#> $prefix
#> [1] "-B="
#> 
#> $separate
#> [1] FALSE
#> 
#> $itemSeparator
#> character(0)
#> 
#> $valueFrom
#> character(0)
#> 
#>