Input parameters InputParam

parameter for a command tool. More details: https://www.commonwl.org/v1.0/CommandLineTool.html#CommandInputParameter

InputParam(
  id,
  label = "",
  type = "string",
  doc = character(),
  secondaryFiles = character(),
  streamable = logical(),
  format = character(),
  loadListing = character(),
  loadContents = logical(),
  position = 0L,
  prefix = "",
  separate = TRUE,
  itemSeparator = character(),
  valueFrom = character(),
  shellQuote = logical(),
  default = character(),
  value = character()
)

# S4 method for cwlProcess
$(x, name)

# S4 method for cwlProcess
$(x, name) <- value

Arguments

id
The unique identifier for this parameter object.

label
A short, human-readable label of this object.

type
valid types of data that may be assigned to this parameter.

doc
Optional. This argument takes an arbitrary documentation as a note for this object.

secondaryFiles
Only valid when type: File or is an array of items: File. Provides a pattern or expression specifying files or directories that must be included alongside the primary file.

streamable
Only valid when type: File or is an array of items: File. A value of true indicates that the file is read or written sequentially without seeking.

format
Only valid when type: File or is an array of items: File.

loadListing
Only valid when type: Directory or is an array of items: Directory.

loadContents
Only valid when type: File or is an array of items: File.

position
The position for this parameter.

prefix
Command line prefix to add before the value.

separate
If true (default), then the prefix and value must be added as separate command line arguments; if false, prefix and value must be concatenated into a single command line argument.

itemSeparator
Join the array elements into a single string with the elements separated by by itemSeparator.

valueFrom
String or Expression.

shellQuote
If ShellCommandRequirement is in the requirements for the current command, this controls whether the value is quoted on the command line (default is true).

default
The default value for this parameter

value
Assigned value for this parameter

x
A `cwlProcess` object.

name
One one of input list

Value

An object of class `InputParam`.

Examples

input1 <- InputParam(id = "sth")