cwlStep function

Constructor function for `cwlStep` object.

cwlStep(
  id,
  run = cwlProcess(),
  In = list(),
  Out = list(),
  scatter = character(),
  scatterMethod = character(),
  label = character(),
  doc = character(),
  requirements = list(),
  hints = list(),
  when = character()
)

Arguments

id
A user-defined unique identifier for this workflow step.

run
A `cwlProcess` object for command line tool, or path to a CWL file.

In
A list of input parameters which will be constructed into `stepInParamList`.

Out
A list of outputs.

scatter
character or a list. The inputs to be scattered.

scatterMethod
required if scatter is an array of more than one element. It can be one of "dotproduct", "nested_crossproduct" and "flat_crossproduct".

label
A short, human-readable label of this object.

doc
A documentation string for this object, or an array of strings which should be concatenated.

requirements
Requirements that apply to either the runtime environment or the workflow engine.

hints
Hints applying to either the runtime environment or the workflow engine.

when
If defined, only run the step when the expression evaluates to true. If false the step is skipped.

Value

An object of class `cwlStep`.

Details

For more details: https://www.commonwl.org/v1.0/Workflow.html#WorkflowStep

See also

Examples

s1 <- cwlStep(id = "s1")