stepParam

A workflow step parameters. More details: https://www.commonwl.org/v1.0/Workflow.html#WorkflowStep

stepParam(
  id,
  run = cwlProcess(),
  In = stepInParamList(),
  Out = list(),
  scatter = character(),
  scatterMethod = character(),
  label = character(),
  doc = character(),
  requirements = list(),
  hints = list(),
  when = character()
)

Arguments

id
The unique identifier for this workflow step.

run
A `cwlProcess` object or the path of a cwl file.

In
A `stepInParamList`.

Out
A list of outputs

scatter
character or a list. The inputs to be scattered.

scatterMethod
required if scatter is an array of more than one element. It can be one of "dotproduct", "nested_crossproduct" and "flat_crossproduct". Details: https://www.commonwl.org/v1.0/Workflow.html#WorkflowStep

label
A short, human-readable label of this object.

doc
A documentation string for this object, or an array of strings which should be concatenated.

requirements
requirements that apply to either the runtime environment or the workflow engine.

hints
hints applying to either the runtime environment or the workflow engine

when
If defined, only run the step when the expression evaluates to true. If false the step is skipped.

Value

An object of class `stepParam`.

Examples

s1 <- stepParam(id = "s1")